requirements.txt
PyAutoGUI~=0.9.54
selenium~=4.24.0
Pillow~=9.0.1